+++ src/Umbraco.Community.Sustainability.TestSite/Views/Home.cshtml
@@ -12,9 +12,8 @@
 
     <script src="https://cdn.tailwindcss.com"></script>
     <script src="https://unpkg.com/@@tgwf/co2@latest/dist/iife/index.js"></script>
-    <script type="module" src="/App_Plugins/Umbraco.Community.Sustainability/js/page-tracker.js"></script>
+    <script type="module" src="/App_Plugins/Umbraco.Community.Sustainability/js/resource-checker.js"></script>
     <meta name="twitter:image" content="https://www.tpximpact.com/media/25wfbyai/tpximpact-social-card.png">
-
 
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" crossorigin href="https://fonts.gstatic.com">
