+++ src/Umbraco.Community.Sustainability/Models/ExternalResourceGroup.cs
@@ -3,7 +3,6 @@ using Umbraco.Community.Sustainability.Extensions;
 
 namespace Umbraco.Community.Sustainability.Models
 {
-
     public class ExternalResourceGroup
     {
         public ResourceGroupType Type { get; set; }
@@ -16,6 +15,17 @@ namespace Umbraco.Community.Sustainability.Models
         {
             Type = type;
             Name = type.GetDisplayName();
+        }
+
+        public static string GetInitiatorType(ResourceGroupType groupType)
+        {
+            return groupType switch
+            {
+                ResourceGroupType.Images => "img",
+                ResourceGroupType.Scripts => "script",
+                ResourceGroupType.Stylesheets => "css",
+                _ => string.Empty,
+            };
         }
     }
