+++ src/Umbraco.Community.Sustainability/Models/SustainabilityResponse.cs
@@ -1,6 +1,6 @@
 namespace Umbraco.Community.Sustainability.Models
 {
-    public class SustainabilityCheckData
+    public class SustainabilityResponse
     {
         public decimal TotalSize { get; set; } = 0;
         public float TotalEmissions { get; set; } = 0;
