+++ src/Umbraco.Community.Sustainability/wwwroot/Umbraco.Community.Sustainability/js/sustainability.resource.js
@@ -1,24 +1,24 @@
 (function () {
-    'use strict';
+  'use strict';
 
-    function sustainabilityResource($http, umbRequestHelper) {
+  function sustainabilityResource($http, umbRequestHelper) {
 
-        var apiUrl = "/umbraco/api/Sustainability/CheckPage";
+    var apiUrl = "/umbraco/api/Sustainability/CheckPage";
 
-        var resource = {
-            getData: getData,
-        };
-       
-        function getData(pageId) {
-            return umbRequestHelper.resourcePromise(
-                $http.get(`${apiUrl}?pageId=${pageId}`),
-                'Failed getting block preview markup'
-            );
-      };
+    var resource = {
+      getData: getData,
+    };
 
-      return resource;
-    }
+    function getData(pageId) {
+      return umbRequestHelper.resourcePromise(
+        $http.get(`${apiUrl}?pageId=${pageId}`),
+        'Failed getting sustainability data'
+      );
+    };
 
-    angular.module('umbraco.resources').factory('Umbraco.Sustainability.Resources.SustainabilityResource', sustainabilityResource);
+    return resource;
+  }
+
+  angular.module('umbraco.resources').factory('Umbraco.Sustainability.Resources.SustainabilityResource', sustainabilityResource);
 
 })();
