+++ src/Umbraco.Community.Sustainability/wwwroot/Umbraco.Community.Sustainability/views/sustainability-check.html
@@ -1,40 +1,40 @@
-<script src="https://unpkg.com/@tgwf/co2@latest/dist/iife/index.js"></script>
-
 <div ng-controller="Umbraco.Sustainability.Controller as vm">
-    <div class="umb-logviewer" ng-if="!loading">
-        <div class="umb-logviewer__main-content">
-            <umb-box ng-repeat="group in sustainabilityData.resourceGroups">
-                <umb-box-header title="{{group.name}}" description="Total size: {{(group.totalSize / 1024).toFixed(2)}}kb"></umb-box-header>
-                <umb-box-content>
-                    <p ng-repeat="resource in group.resources">
-                        {{resource.url}} ({{(resource.size / 1024).toFixed(2)}}kb)
-                    </p>
-                </umb-box-content>
-            </umb-box>
-        </div>
-        <div class="umb-logviewer__sidebar">
-            <div class="umb-box-row">
-                <umb-box>
-                    <umb-box-header title="Page size"></umb-box-header>
-                    <umb-box-content>{{(sustainabilityData.totalSize / 1024).toFixed(2)}}kb</umb-box-content>
-                </umb-box>
-                <umb-box>
-                    <umb-box-header title="CO₂ per page view"></umb-box-header>
-                    <umb-box-content>{{sustainabilityData.totalEmissions.toFixed(4)}}g</umb-box-content>
-                </umb-box>
-            </div>
-
-            <umb-box>
-                <umb-box-header title="Breakdown by type"></umb-box-header>
-                <umb-box-content class="block-form">
-                    <canvas chart-doughnut
-                            chart-data="vm.datasets"
-                            chart-labels="vm.labels"
-                            chart-colours="vm.colours"
-                            chart-options="vm.options">
-                    </canvas>
-                </umb-box-content>
-            </umb-box>
-        </div>
+  <div class="umb-logviewer" ng-if="!loading">
+    <div class="umb-logviewer__main-content">
+      <umb-box ng-repeat="group in sustainabilityData.resourceGroups">
+        <umb-box-header title="{{group.name}}" description="Total size: {{(group.totalSize / 1024).toFixed(2)}}kb"></umb-box-header>
+        <umb-box-content>
+          <ul>
+            <li ng-repeat="resource in group.resources">
+              {{resource.url}} ({{(resource.size / 1024).toFixed(2)}}kb)
+            </li>
+          </ul>
+        </umb-box-content>
+      </umb-box>
     </div>
+    <div class="umb-logviewer__sidebar">
+      <div class="umb-box-row">
+        <umb-box>
+          <umb-box-header title="Page size"></umb-box-header>
+          <umb-box-content>{{(sustainabilityData.totalSize / 1024).toFixed(2)}}kb</umb-box-content>
+        </umb-box>
+        <umb-box>
+          <umb-box-header title="CO₂ per page view"></umb-box-header>
+          <umb-box-content>{{sustainabilityData.totalEmissions.toFixed(4)}}g</umb-box-content>
+        </umb-box>
+      </div>
+
+      <umb-box>
+        <umb-box-header title="Breakdown by type"></umb-box-header>
+        <umb-box-content class="block-form">
+          <canvas chart-doughnut
+                  chart-data="vm.datasets"
+                  chart-labels="vm.labels"
+                  chart-colours="vm.colours"
+                  chart-options="vm.options">
+          </canvas>
+        </umb-box-content>
+      </umb-box>
+    </div>
+  </div>
 </div>
